chore: function logic split into multiple files
