+++ functions/config.py
@@ -1,0 +1,5 @@
+from firebase_admin import initialize_app
+
+
+def initialize_firebase():
+    initialize_app()
